arm64: mm: dma-mapping: fix size type to support >= 4GB size.
Bug 1860962

Reviewed-on: https://git-master.nvidia.com/r/1541427
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/ata/libata-sff.c
@@ -853,7 +853,7 @@ next_sg:
 	offset %= PAGE_SIZE;
 
 	/* don't overrun current sg */
-	count = min(sg->length - qc->cursg_ofs, bytes);
+	count = min(sg->length - qc->cursg_ofs, (size_t)bytes);
 
 	/* don't cross page boundaries */
 	count = min(count, (unsigned int)PAGE_SIZE - offset);
